Пометка для новых сообщений. Улучшения

+++ backend/Backend/src/main/java/ru/nikidzawa/backend/store/entity/ChatEntity.java
@@ -27,7 +27,7 @@ public class ChatEntity {
     ChatType type;
 
     @Column(name = "last_message_id")
-    Integer lastMessageId;
+    Long lastMessageId;
 
     @Column(name = "created_at")
     LocalDateTime createdAt;
